Shards anim remove gfx dependency
