+++ shards/gfx/error_utils.hpp
@@ -1,13 +1,13 @@
 #ifndef GFX_ERROR_UTILS
 #define GFX_ERROR_UTILS
 
+#include <shards/core/exception.hpp>
 #include <spdlog/fmt/fmt.h>
 #include <stdexcept>
 
 namespace gfx {
-template <typename... TArgs> 
-std::runtime_error formatException(fmt::format_string<TArgs...> format, TArgs&&... args) {
-  return std::runtime_error(fmt::format(format, std::forward<TArgs>(args)...));
+template <typename... TArgs> std::runtime_error formatException(fmt::format_string<TArgs...> format, TArgs &&...args) {
+  return shards::formatException(format, std::forward<TArgs>(args)...);
 }
 } // namespace gfx
