+++ shards/modules/anim/anim.cpp
@@ -1,11 +1,11 @@
-#include "anim/path.hpp"
-#include "anim/types.hpp"
+#include <shards/anim/path.hpp>
+#include <shards/anim/types.hpp>
 #include <shards/core/shared.hpp>
 #include "linalg.h"
 #include <shards/linalg_shim.hpp>
 #include <shards/modules/core/math.hpp>
 #include <shards/modules/core/time.hpp>
-#include <shards/modules/gfx/shards_utils.hpp>
+#include <shards/core/check_utils.hpp>
 #include <shards/math_ops.hpp>
 #include <shards/shardwrapper.hpp>
 #include <shards/core/params.hpp>
@@ -24,7 +24,7 @@ static auto getKeyframeValue(const SHVar &keyframe) { return ((TableVar &)keyfra
 static auto getKeyframeInterpolation(const SHVar &keyframe) {
   Var &v = ((TableVar &)keyframe).get<Var>(Var("Interpolation"));
   if (v.valueType == SHType::Enum) {
-    gfx::checkEnumType(v, ShardsTypes::InterpolationEnumInfo::Type, "Interpolation");
+    shards::checkEnumType(v, ShardsTypes::InterpolationEnumInfo::Type, "Interpolation");
     return (Interpolation)v.payload.enumValue;
   }
   return Interpolation::Linear;
